Add a decode benchmark
